pyproject.toml
[tool.poetry]
name = "practice_bq"
version = "0.1.0"
description = ""
authors = ["o-r-r-p <ryosuke.otomo@rakus-partners.co.jp>"]

[tool.poetry.dependencies]
python = ">=3.7,<3.11"
google-cloud-bigquery = "^3.1.0"
google-cloud-storage = "^2.3.0"
pdfminer = "^20191125"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
